// frontend/apps/frontend/src/app/polls-backend/model/pollDto.ts
/**
 * CorePlugin.BackendDevServer
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PollOptionDto } from './pollOptionDto';


export interface PollDto {
    pollCode: string;
    pollName: string;
    creatorName: string;
    startTime: string;
    endTime: string;
    isMultipleChoice: boolean;
    pollOptions: Array<PollOptionDto>;
    pollQuestion: string;
}